Remove `Backtrace` class and cleanup related code
Eliminate `Backtrace` class, associated methods, and header file. Adjust `main.cpp` to remove periodic backtrace reporting. Simplify logging in `Watcher::setStandby`. Migrate relevant debugging functionality to `Guardian` with new shutdown and exception handlers. Update `platformio.ini` to improve debugging configuration and refine documentation in README.

+++ src/Fader.cpp
@@ -6,7 +6,7 @@
 #include "Fader.h"
 
 bool allowFade = false;
-uint8_t fadeSpeed = 500;
+uint64_t fadeSpeed = 500;
 uint8_t fadeDuty = 255;
 uint8_t currentDuty = 0;
 unsigned long lastMs = 0;
